requirements.txt
# Procesamiento de imágenes
opencv-python==4.8.0.76
pytesseract==0.3.10
pillow==10.0.0
imutils==0.5.4

# Análisis de datos
pandas==2.0.3
numpy==1.24.3

# Utilidades
python-dotenv==1.0.0
tqdm==4.65.0